Update test modules
